+++ src/btc/btc_client.rs
@@ -20,10 +20,10 @@ impl BlockchainClient for BtcClient {
 
     fn get_block_with_tx_count_for_height(
         &self,
-        height: u64,
+        height: u32,
     ) -> Result<(bitcoin::Block, usize), String> {
         self.rpc_client
-            .get_block_hash(height)
+            .get_block_hash(height as u64)
             .map_err(|e| e.to_string())
             .and_then(|hash| {
                 self.get_block_with_tx_count(&hash)
